+++ packages/form/src/components/List/index.tsx
@@ -25,15 +25,19 @@ const FormListContext = React.createContext<
   | Record<string, any>
 >({});
 
-type ChildrenFunction = (
-  fields: FormListFieldData[],
-  operation: FormListOperation,
-  meta: {
-    errors: React.ReactNode[];
-  },
-) => React.ReactNode;
+// type ChildrenFunction = (
+//   fields: FormListFieldData[],
+//   operation: FormListOperation,
+//   meta: {
+//     errors: React.ReactNode[];
+//   },
+// ) => React.ReactNode;
 
-type ChildrenItemFunction = (field: FormListFieldData, index: number) => React.ReactNode;
+type ChildrenItemFunction = (
+  field: FormListFieldData,
+  index: number,
+  operation: FormListOperation,
+) => React.ReactNode;
 
 export type ProFormListProps = Omit<FormListProps, 'children'> & {
   creatorButtonProps?:
@@ -50,7 +54,7 @@ export type ProFormListProps = Omit<FormListProps, 'children'> & {
     action: FormListOperation,
     defaultActionDom: ReactNode[],
   ) => ReactNode[];
-  children: ReactNode | ChildrenFunction;
+  children: ReactNode | ChildrenItemFunction;
   itemContainerRender?: (
     doms: ReactNode,
     listMeta: {
@@ -96,7 +100,6 @@ const listToArray = (children?: ReactNode | ReactNode[]) => {
 
 type ProFormListItemProps = {
   creatorButtonProps: ProFormListProps['creatorButtonProps'];
-
   formInstance: FormInstance;
   copyIconProps: ProFormListProps['copyIconProps'];
   deleteIconProps: ProFormListProps['deleteIconProps'];
@@ -111,6 +114,7 @@ type ProFormListItemProps = {
     errors: ReactNode[];
   };
   name: ProFormListProps['name'];
+  originName: ProFormListProps['name'];
 };
 
 const ProFormListItem: React.FC<
@@ -142,7 +146,7 @@ const ProFormListItem: React.FC<
   const childrenArray = listToArray(children)
     .map((childrenItem) => {
       if (typeof childrenItem === 'function') {
-        return (childrenItem as ChildrenItemFunction)?.(field, index);
+        return (childrenItem as ChildrenItemFunction)?.(field, index, action);
       }
       return childrenItem;
     })
@@ -228,13 +232,12 @@ const ProFormListItem: React.FC<
       {dom}
     </div>
   );
-
   return (
     <FormListContext.Provider
       key={field.name}
       value={{
         ...field,
-        listName: [listContext.listName, rest.name, field.name]
+        listName: [listContext.listName, rest.originName, field.name]
           .filter((item) => item !== undefined)
           .flat(1),
       }}
@@ -344,6 +347,7 @@ const ProFormList: React.FC<ProFormListProps> = ({
                   <>
                     <ProFormListContainer
                       name={name}
+                      originName={rest.name}
                       copyIconProps={copyIconProps}
                       deleteIconProps={deleteIconProps}
                       formInstance={formInstance as any}