+++ packages/provider/src/locale/fa_IR.tsx
@@ -53,6 +53,9 @@ export default {
     prev: 'قبلی',
     submit: 'اتمام',
   },
+  loginForm: {
+    submitText: 'ورود',
+  },
   editableTable: {
     action: {
       save: 'ذخیره',