+++ packages/provider/src/locale/ms_MY.tsx
@@ -40,4 +40,7 @@ export default {
     densityMiddle: 'Tengah',
     densitySmall: 'Kecil',
   },
+  loginForm: {
+    submitText: 'Log Masuk',
+  },
 };