+++ packages/card/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.15.3...@ant-design/pro-card@1.15.4) (2021-10-15)
+
+**Note:** Version bump only for package @ant-design/pro-card
+
 ## [1.15.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-card@1.15.2...@ant-design/pro-card@1.15.3) (2021-10-12)
 
 ### Bug Fixes