+++ packages/provider/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.20](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@1.4.19...@ant-design/pro-provider@1.4.20) (2021-10-15)
+
+**Note:** Version bump only for package @ant-design/pro-provider
+
 ## [1.4.19](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@1.4.18...@ant-design/pro-provider@1.4.19) (2021-09-24)
 
 ### Bug Fixes